requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
attrs==24.3.0
ccxt==4.4.49
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pyTelegramBotAPI==4.26.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
